Review workflows test coverage
Closes #42694
