+++ tests/base/src/test/java/org/keycloak/tests/workflow/activation/UserCreationWorkflowTest.java
@@ -1,0 +1,61 @@
+package org.keycloak.tests.workflow.activation;
+
+import jakarta.ws.rs.core.Response;
+
+import org.keycloak.admin.client.resource.UserResource;
+import org.keycloak.models.workflow.ResourceOperationType;
+import org.keycloak.models.workflow.SetUserAttributeStepProviderFactory;
+import org.keycloak.representations.idm.UserRepresentation;
+import org.keycloak.representations.userprofile.config.UPConfig;
+import org.keycloak.representations.workflows.WorkflowRepresentation;
+import org.keycloak.representations.workflows.WorkflowStepRepresentation;
+import org.keycloak.testframework.annotations.KeycloakIntegrationTest;
+import org.keycloak.testframework.realm.UserConfigBuilder;
+import org.keycloak.testframework.util.ApiUtil;
+import org.keycloak.tests.workflow.AbstractWorkflowTest;
+import org.keycloak.tests.workflow.config.WorkflowsBlockingServerConfig;
+
+import org.junit.jupiter.api.Test;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.notNullValue;
+
+/**
+ * Tests activation of workflows based on user creation events.
+ */
+@KeycloakIntegrationTest(config = WorkflowsBlockingServerConfig.class)
+public class UserCreationWorkflowTest extends AbstractWorkflowTest {
+
+    @Test
+    public void testActivateWorkflowOnUserCreation() {
+        UPConfig upConfig = managedRealm.admin().users().userProfile().getConfiguration();
+        upConfig.setUnmanagedAttributePolicy(UPConfig.UnmanagedAttributePolicy.ENABLED);
+        managedRealm.admin().users().userProfile().update(upConfig);
+
+        // create the workflow that triggers on user creation
+        WorkflowRepresentation workflow = WorkflowRepresentation.withName("myworkflow")
+                .onEvent(ResourceOperationType.USER_CREATED.name())
+                .withSteps(
+                        WorkflowStepRepresentation.create()
+                                .of(SetUserAttributeStepProviderFactory.ID)
+                                .withConfig("attribute", "attr1")
+                                .build()
+                ).build();
+        managedRealm.admin().workflows().create(workflow).close();
+
+        // create a test user to trigger the workflow
+        String userId;
+        try (Response response = managedRealm.admin().users().create(UserConfigBuilder.create()
+                .username("generic-user").email("generic-user@example.com").build())) {
+            userId = ApiUtil.getCreatedId(response);
+        }
+        UserResource userResource = managedRealm.admin().users().get(userId);
+        UserRepresentation representation = userResource.toRepresentation();
+
+        // assert that the workflow step ran and set the user attribute
+        assertThat(representation.getAttributes(), notNullValue());
+        assertThat(representation.getAttributes().get("attribute"), notNullValue());
+        assertThat(representation.getAttributes().get("attribute").get(0), is("attr1"));
+    }
+}
