+++ tests/base/src/test/java/org/keycloak/tests/workflow/config/WorkflowsBlockingServerConfig.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.keycloak.tests.workflow;
+package org.keycloak.tests.workflow.config;
 
 import org.keycloak.testframework.server.KeycloakServerConfigBuilder;
