+++ tests/base/src/test/java/org/keycloak/tests/workflow/execution/BrokeredUserLifecycleWorkflowTest.java
@@ -1,37 +1,19 @@
-/*
- * Copyright 2025 Red Hat, Inc. and/or its affiliates
- * and other contributors as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.keycloak.tests.workflow;
+package org.keycloak.tests.workflow.execution;
 
 import java.time.Duration;
 import java.util.List;
 import java.util.Map;
 
+import jakarta.mail.internet.MimeMessage;
 import jakarta.ws.rs.core.Response;
 
 import org.keycloak.admin.client.resource.UsersResource;
 import org.keycloak.models.RealmModel;
 import org.keycloak.models.UserModel;
 import org.keycloak.models.workflow.DeleteUserStepProviderFactory;
+import org.keycloak.models.workflow.DisableUserStepProviderFactory;
+import org.keycloak.models.workflow.NotifyUserStepProviderFactory;
 import org.keycloak.models.workflow.ResourceOperationType;
-import org.keycloak.models.workflow.Workflow;
-import org.keycloak.models.workflow.WorkflowProvider;
-import org.keycloak.models.workflow.WorkflowStateProvider;
-import org.keycloak.models.workflow.WorkflowStateProvider.ScheduledStep;
 import org.keycloak.models.workflow.conditions.IdentityProviderWorkflowConditionFactory;
 import org.keycloak.representations.idm.FederatedIdentityRepresentation;
 import org.keycloak.representations.idm.IdentityProviderRepresentation;
@@ -44,6 +26,8 @@ import org.keycloak.testframework.annotations.InjectRealm;
 import org.keycloak.testframework.annotations.InjectUser;
 import org.keycloak.testframework.annotations.KeycloakIntegrationTest;
 import org.keycloak.testframework.injection.LifeCycle;
+import org.keycloak.testframework.mail.MailServer;
+import org.keycloak.testframework.mail.annotations.InjectMailServer;
 import org.keycloak.testframework.oauth.OAuthClient;
 import org.keycloak.testframework.oauth.annotations.InjectOAuthClient;
 import org.keycloak.testframework.realm.ClientConfig;
@@ -63,18 +47,22 @@ import org.keycloak.testframework.ui.page.ConsentPage;
 import org.keycloak.testframework.ui.page.LoginPage;
 import org.keycloak.testframework.ui.webdriver.ManagedWebDriver;
 import org.keycloak.testframework.util.ApiUtil;
+import org.keycloak.tests.workflow.AbstractWorkflowTest;
+import org.keycloak.tests.workflow.config.WorkflowsBlockingServerConfig;
 
 import org.awaitility.Awaitility;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import static org.keycloak.models.workflow.ResourceOperationType.USER_AUTHENTICATED;
 import static org.keycloak.models.workflow.ResourceOperationType.USER_CREATED;
+import static org.keycloak.tests.workflow.util.EmailTestUtils.findEmailByRecipient;
+import static org.keycloak.tests.workflow.util.EmailTestUtils.verifyEmailContent;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.containsString;
+import static org.hamcrest.Matchers.empty;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
@@ -85,9 +73,11 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
+ * Tests the usage of workflows to manage the lifecycle of brokered users. The idea is to track the user's activity through
+ * user-authenticated events, and deactivate or even delete the user after a certain period of inactivity.
  */
 @KeycloakIntegrationTest(config = WorkflowsBlockingServerConfig.class)
-public class BrokeredUserSessionRefreshTimeWorkflowTest extends AbstractWorkflowTest {
+public class BrokeredUserLifecycleWorkflowTest extends AbstractWorkflowTest {
 
     @InjectRunOnServer(permittedPackages = "org.keycloak.tests", realmRef = "consumer")
     RunOnServerClient runOnServer;
@@ -119,6 +109,9 @@ public class BrokeredUserSessionRefreshTimeWorkflowTest extends AbstractWorkflow
     @InjectPage
     ConsentPage consentPage;
 
+    @InjectMailServer
+    private MailServer mailServer;
+
     private static final String REALM_PROV_NAME = "provider";
     private static final String REALM_CONS_NAME = "consumer";
 
@@ -130,6 +123,100 @@ public class BrokeredUserSessionRefreshTimeWorkflowTest extends AbstractWorkflow
     private static final String CLIENT_SECRET = "secret";
 
     @Test
+    public void testWorkflowToManageBrokeredUserLifecycle() {
+
+        // create a workflow that notifies inactive users after 7 days, disables them 30 days after that if the user doesn't
+        // log back in, and finally deletes them also 30 days after being disabled.
+        consumerRealm.admin().workflows().create(WorkflowRepresentation.withName("myworkflow")
+                .onEvent(USER_AUTHENTICATED.toString())
+                .onCondition(IDP_CONDITION)
+                .concurrency().restartInProgress("true")
+                .withSteps(
+                        WorkflowStepRepresentation.create().of(NotifyUserStepProviderFactory.ID)
+                                .after(Duration.ofDays(3))
+                                .withConfig("reason", "inactivity")
+                                .build(),
+                        WorkflowStepRepresentation.create().of(DisableUserStepProviderFactory.ID)
+                                .after(Duration.ofDays(10))
+                                .build(),
+                        WorkflowStepRepresentation.create().of(DeleteUserStepProviderFactory.ID)
+                                .after(Duration.ofDays(10))
+                                .build()
+                ).build()).close();
+
+        // login as alice - this should trigger the workflow execution for her user
+        loginBrokeredUser();
+
+        UsersResource users = consumerRealm.admin().users();
+        String username = aliceFromProviderRealm.getUsername();
+        UserRepresentation federatedUser = users.search(username).get(0);
+        List<FederatedIdentityRepresentation> federatedIdentities = users.get(federatedUser.getId()).getFederatedIdentity();
+        assertFalse(federatedIdentities.isEmpty());
+
+        // check that alice is associated with the workflow
+        assertScheduledWorkflows(federatedUser.getId(), NotifyUserStepProviderFactory.ID, 3);
+
+        // simulate 8 days passing - alice should be notified but still enabled
+        runScheduledSteps(Duration.ofDays(4));
+
+        MimeMessage testUserMessage = findEmailByRecipient(mailServer, "alice@wonderland.org");
+        assertNotNull(testUserMessage, "No email found for alice@wonderland.org");
+        verifyEmailContent(testUserMessage, "alice@wonderland.org", "Disable", "Alice", "10", "inactivity");
+
+        federatedUser = users.get(federatedUser.getId()).toRepresentation();
+        assertThat(federatedUser.isEnabled(), is(true));
+        assertScheduledWorkflows(federatedUser.getId(), DisableUserStepProviderFactory.ID, 2); // next scheduled step is disable
+
+        // advance more time - alice should now be disabled
+        runScheduledSteps(Duration.ofDays(15));
+        federatedUser = users.get(federatedUser.getId()).toRepresentation();
+        assertThat(federatedUser.isEnabled(), is(false));
+        assertScheduledWorkflows(federatedUser.getId(), DeleteUserStepProviderFactory.ID, 1);
+
+        // let's re-enable alice and then login again to simulate her coming back
+        federatedUser.setEnabled(true);
+        users.get(federatedUser.getId()).update(federatedUser);
+        consumerRealmOAuth.openLoginForm();
+
+        // workflow should have been restarted due to reauthentication
+        assertScheduledWorkflows(federatedUser.getId(), NotifyUserStepProviderFactory.ID, 3);
+
+        // now simulate the workflow to its end - alice should be deleted
+        runScheduledSteps(Duration.ofDays(5)); // notify
+        runScheduledSteps(Duration.ofDays(10)); // disable
+        runScheduledSteps(Duration.ofDays(10)); // delete
+
+        runOnServer.run((session -> {
+            RealmModel realm = session.getContext().getRealm();
+            UserModel user = session.users().getUserByUsername(realm, username);
+            assertNull(user);
+        }));
+    }
+
+    @Test
+    public void testNonBrokeredUserNotAffectedByWorkflow() {
+        // create a workflow that deletes inactive users after 10 days.
+        consumerRealm.admin().workflows().create(WorkflowRepresentation.withName("myworkflow")
+                .onEvent(USER_AUTHENTICATED.toString())
+                .onCondition(IDP_CONDITION)
+                .withSteps(
+                        WorkflowStepRepresentation.create().of(DeleteUserStepProviderFactory.ID)
+                                .after(Duration.ofDays(10))
+                                .build()
+                ).build()).close();
+
+        // authenticate with bob directly in the consumer realm - he is not associated with the IDP and thus not influenced
+        // by the idp-exclusive lifecycle workflow.
+        consumerRealmOAuth.openLoginForm();
+        loginPage.fillLogin(bobFromConsumerRealm.getUsername(), bobFromConsumerRealm.getPassword());
+        loginPage.submit();
+        assertTrue(driver.page().getPageSource().contains("Happy days"), "Test user should be successfully logged in.");
+
+        List<WorkflowRepresentation> scheduledWorkflows = consumerRealm.admin().workflows().getScheduledWorkflows(bobFromConsumerRealm.getId());
+        assertThat(scheduledWorkflows, hasSize(0));
+    }
+
+    @Test
     public void testInvalidateWorkflowOnIdentityProviderRemoval() {
         String workflowId;
         try (Response response = consumerRealm.admin().workflows().create(WorkflowRepresentation.withName("myworkflow")
@@ -137,8 +224,8 @@ public class BrokeredUserSessionRefreshTimeWorkflowTest extends AbstractWorkflow
                 .onCondition(IDP_CONDITION)
                 .withSteps(
                         WorkflowStepRepresentation.create().of(DeleteUserStepProviderFactory.ID)
-                            .after(Duration.ofDays(1))
-                            .build())
+                                .after(Duration.ofDays(1))
+                                .build())
                 .build())) {
             workflowId = ApiUtil.getCreatedId(response);
         }
@@ -168,11 +255,18 @@ public class BrokeredUserSessionRefreshTimeWorkflowTest extends AbstractWorkflow
                 });
     }
 
+    /**
+     * Tests a somewhat different workflow that activates when a user links a federated identity from a specific IDP, and
+     * that is cancelled if the identity is removed. The idea is to delete users that link that specific IDP after a certain period
+     * of time unless they unlink the IDP before that.
+     */
     @Test
-    public void tesRunStepOnFederatedUser() {
+    public void testWorkflowBasedOnFederatedIdentityMembership() {
+
+        // create a workflow that deletes users 1 day after a federated identity is added, and that is cancelled if the identity is removed
         consumerRealm.admin().workflows().create(WorkflowRepresentation.withName("myworkflow")
-                .onEvent(USER_CREATED.toString(), USER_AUTHENTICATED.toString())
-                .onCondition(IDP_CONDITION)
+                .onEvent(ResourceOperationType.USER_FEDERATED_IDENTITY_ADDED.name() + "(" + IDP_OIDC_ALIAS + ")")
+                .concurrency().cancelInProgress(ResourceOperationType.USER_FEDERATED_IDENTITY_REMOVED.name() + "(" + IDP_OIDC_ALIAS + ")")
                 .withSteps(
                         WorkflowStepRepresentation.create().of(DeleteUserStepProviderFactory.ID)
                                 .after(Duration.ofDays(1))
@@ -181,90 +275,16 @@ public class BrokeredUserSessionRefreshTimeWorkflowTest extends AbstractWorkflow
 
         loginBrokeredUser();
 
-        UsersResource users = consumerRealm.admin().users();
-        String username = aliceFromProviderRealm.getUsername();
-        UserRepresentation federatedUser = users.search(username).get(0);
-        List<FederatedIdentityRepresentation> federatedIdentities = users.get(federatedUser.getId()).getFederatedIdentity();
-        assertFalse(federatedIdentities.isEmpty());
-
-        runScheduledSteps(Duration.ZERO);
-
-        runOnServer.run((session -> {
-            RealmModel realm = session.getContext().getRealm();
-            UserModel user = session.users().getUserByUsername(realm, username);
-            assertNotNull(user);
-            assertTrue(user.isEnabled());
-        }));
-
-        runScheduledSteps(Duration.ofDays(2));
-
-        runOnServer.run((session -> {
-            RealmModel realm = session.getContext().getRealm();
-            UserModel user = session.users().getUserByUsername(realm, username);
-            assertNull(user);
-        }));
-
-        // now authenticate with bob directly in the consumer realm - he is not associated with the IDP and thus not influenced
-        // by the idp-exclusive lifecycle workflow.
-        consumerRealmOAuth.openLoginForm();
-        loginPage.fillLogin(bobFromConsumerRealm.getUsername(), bobFromConsumerRealm.getPassword());
-        loginPage.submit();
-        assertTrue(driver.page().getPageSource().contains("Happy days"), "Test user should be successfully logged in.");
-
-        // run the scheduled tasks - bob should not be affected.
-        runScheduledSteps(Duration.ZERO);
-
-        runOnServer.run(session -> {
-            RealmModel realm = session.getContext().getRealm();
-            UserModel user = session.users().getUserByUsername(realm, "bob");
-            assertNotNull(user);
-            assertTrue(user.isEnabled());
-        });
-
-        // run with a time offset - bob should still not be affected.
-        runScheduledSteps(Duration.ofDays(2));
-
-        runOnServer.run(session -> {
-            RealmModel realm = session.getContext().getRealm();
-            UserModel user = session.users().getUserByUsername(realm, "bob");
-            assertNotNull(user);
-        });
-    }
-
-    @Disabled("For now deactivation events is not enabled to any event")
-    @Test
-    public void testAddRemoveFedIdentityAffectsWorkflowAssociation() {
-        consumerRealm.admin().workflows().create(WorkflowRepresentation.withName("myworkflow")
-                .onEvent(ResourceOperationType.USER_FEDERATED_IDENTITY_ADDED.toString())
-                .onCondition(IDP_CONDITION)
-                .withSteps(
-                        WorkflowStepRepresentation.create().of(DeleteUserStepProviderFactory.ID)
-                                .after(Duration.ofDays(1))
-                                .build()
-                ).build()).close();
-
-        loginBrokeredUser();
-
-        runOnServer.run(session -> {
-            RealmModel realm = session.getContext().getRealm();
-            WorkflowProvider provider = session.getProvider(WorkflowProvider.class);
-            Workflow workflow = provider.getWorkflows().toList().get(0);
-            UserModel alice = session.users().getUserByUsername(realm, "alice");
-            assertNotNull(alice);
-
-            // alice should be associated with the workflow
-            WorkflowStateProvider stateProvider = session.getProvider(WorkflowStateProvider.class);
-            ScheduledStep scheduledStep = stateProvider.getScheduledStep(workflow.getId(), alice.getId());
-            assertNotNull(scheduledStep, "A step should have been scheduled for the user " + alice.getUsername());
-        });
-
-        // remove the federated identity - alice should be disassociated from the workflow and thus not deleted
         UserRepresentation aliceInConsumerRealm = consumerRealm.admin().users().search(aliceFromProviderRealm.getUsername()).get(0);
         assertNotNull(aliceInConsumerRealm);
+        assertScheduledWorkflows(aliceInConsumerRealm.getId(), DeleteUserStepProviderFactory.ID, 1);
+
+        // remove the federated identity - alice should be disassociated from the workflow and thus not deleted
         consumerRealm.admin().users().get(aliceInConsumerRealm.getId()).removeFederatedIdentity(IDP_OIDC_ALIAS);
 
         // run with a time offset - alice should not be deleted as she is no longer associated with the IDP and thus the workflow
-        runScheduledSteps(Duration.ofDays(2));
+        List<WorkflowRepresentation> workflows = consumerRealm.admin().workflows().getScheduledWorkflows(aliceInConsumerRealm.getId());
+        assertThat(workflows, empty());
 
         runOnServer.run(session -> {
             RealmModel realm = session.getContext().getRealm();
@@ -288,6 +308,17 @@ public class BrokeredUserSessionRefreshTimeWorkflowTest extends AbstractWorkflow
             assertNull(user);
         });
 
+    }
+
+    private void assertScheduledWorkflows(String resourceId, String nextScheduledStep, int expectedSteps) {
+        List<WorkflowRepresentation> scheduledWorkflows = consumerRealm.admin().workflows().getScheduledWorkflows(resourceId);
+        assertThat(scheduledWorkflows, hasSize(1));
+        WorkflowRepresentation scheduledWorkflow = scheduledWorkflows.get(0);
+        assertThat(scheduledWorkflow.getName(), is("myworkflow"));
+        assertThat(scheduledWorkflow.getSteps(), hasSize(expectedSteps));
+        if (nextScheduledStep != null) {
+            assertThat(scheduledWorkflow.getSteps().get(0).getUses(), is(nextScheduledStep));
+        }
     }
 
     private void loginBrokeredUser() {
